\name{collinear}
\alias{collinear}
\title{
Three Points Collinear
}
\description{
  Assess whether the three points are collinear.
}
\usage{
  collinear(da,deviation=0.01)
}

\arguments{
  \item{da}{
    a N x 2 matrix or data frame (rows: N points; columns: x coordinate value,y coordinate value) with one header c('x','y').
  }
  \item{deviation}{
   Slope difference, used to evaluate error terms for three-point collinear.
}
}

\details{
  
}

\author{
%%  ~~who you are~~
Wen Luo(design, implementation, testing)
}

\examples{

}